us_15 test cases updated.
