--- src/test/java/allover/tests/US_15/TC_01_DisplayMenu.java
@@ -1,4 +1,0 @@
-package allover.tests.US_15;
-
-public class TC_01_DisplayMenu {
-}
